python/requirements.txt
JPype1==1.0.1
seldon-core==1.2.1
numpy==1.18.5
